Refactor user authentication and game routes; enhance Google login handling and database connection setup
