+++ server/app.js
@@ -4,6 +4,7 @@ const cors = require("cors");
 const { Server } = require("socket.io");
 require('dotenv').config();
 
+
 const authRoutes = require("./routes/authRoutes");
 const userRoutes = require("./routes/userRoutes");
 const gameRoutes = require("./routes/gameRoutes");
@@ -23,8 +24,8 @@ app.use(express.json());
 
 // Routes
 app.use('/api/auth', authRoutes);
-app.use('/user', userRoutes);  // this route handles user-related actions and needs to be used in frontend
-app.use('/game', gameRoutes); 
+app.use('/api/user', userRoutes);  // this route handles user-related actions and needs to be used in frontend
+app.use('/api/game', gameRoutes);
 
 // Health check
 app.get('/health', (req, res) => {
