+++ server/models/Problem.js
@@ -1,5 +1,5 @@
 const { DataTypes } = require('sequelize');
-const { sequelize } = require('../config/database');
+const sequelize = require('../models/database');
 
 const Problem = sequelize.define('Problem', {
   id: {
